+++ src/aks-preview/azext_aks_preview/_client_factory.py
@@ -22,11 +22,15 @@ def cf_container_services(cli_ctx, *_):
 
 
 def get_container_service_client(cli_ctx, **_):
-    return get_mgmt_service_client(cli_ctx, CUSTOM_MGMT_AKS)
+    return get_mgmt_service_client(cli_ctx, CUSTOM_MGMT_AKS_PREVIEW)
 
 
 def cf_managed_clusters(cli_ctx, *_):
-    return get_mgmt_service_client(cli_ctx, CUSTOM_MGMT_AKS_PREVIEW)
+    return get_mgmt_service_client(cli_ctx, CUSTOM_MGMT_AKS_PREVIEW).managed_clusters
+
+
+def cf_agent_pools(cli_ctx, *_):
+    return get_mgmt_service_client(cli_ctx, CUSTOM_MGMT_AKS_PREVIEW).agent_pools
 
 
 def cf_resource_groups(cli_ctx, subscription_id=None):