+++ src/aks-preview/azext_aks_preview/tests/latest/test_aks_command_preview.py
@@ -22,6 +22,7 @@ class AzureKubernetesServicePreviewScenarioTest(ScenarioTest):
 
     @ResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='eastus')
     @RoleBasedServicePrincipalPreparer()
+    @AllowLargeResponse
     def test_aks_create_default_service_preview(self, resource_group, resource_group_location, sp_name, sp_password):
         # kwargs for string formatting
         aks_name = self.create_random_name('cliakstest', 16)
@@ -76,6 +77,7 @@ class AzureKubernetesServicePreviewScenarioTest(ScenarioTest):
 
     @ResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='eastus')
     @RoleBasedServicePrincipalPreparer()
+    @AllowLargeResponse
     def test_aks_create_default_service_enable_autoscaler(self, resource_group, resource_group_location, sp_name, sp_password):
         aks_name = self.create_random_name('cliakstest', 16)
         self.kwargs.update({
@@ -90,7 +92,8 @@ class AzureKubernetesServicePreviewScenarioTest(ScenarioTest):
                      '--node-count=2 --service-principal={service_principal} ' \
                      '--client-secret={client_secret} --enable-cluster-autoscaler ' \
                      '--min-count=1 --max-count=3 ' \
-                     '--no-ssh-key'
+                     '--no-ssh-key' \
+                     '--enable-vmss'
         self.cmd(create_cmd, checks=[
             self.check('agentPoolProfiles[0].minCount', '1'),
             self.check('agentPoolProfiles[0].maxCount', '3'),
@@ -100,7 +103,13 @@ class AzureKubernetesServicePreviewScenarioTest(ScenarioTest):
 
         # delete
         self.cmd('aks delete -g {resource_group} -n {name} --yes --no-wait', checks=[self.is_empty()])
-    
+
+    def test_aks_get_versions(self):
+        # show k8s versions
+        self.cmd('aks get-versions -l {location}', checks=[
+            self.exists('orchestrators[*].orchestratorVersion')
+        ])
+
     @classmethod
     def generate_ssh_keys(cls):
         TEST_SSH_KEY_PUB = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQCbIg1guRHbI0lV11wWDt1r2cUdcNd27CJsg+SfgC7miZeubtwUhbsPdhMQsfDyhOWHq1+ZL0M+nJZV63d/1dhmhtgyOqejUwrPlzKhydsbrsdUor+JmNJDdW01v7BXHyuymT8G4s09jCasNOwiufbP/qp72ruu0bIA1nySsvlf9pCQAuFkAnVnf/rFhUlOkhtRpwcq8SUNY2zRHR/EKb/4NWY1JzR4sa3q2fWIJdrrX0DvLoa5g9bIEd4Df79ba7v+yiUBOS0zT2ll+z4g9izHK3EO5d8hL4jYxcjKs+wcslSYRWrascfscLgMlMGh0CdKeNTDjHpGPncaf3Z+FwwwjWeuiNBxv7bJo13/8B/098KlVDl4GZqsoBCEjPyJfV6hO0y/LkRGkk7oHWKgeWAfKtfLItRp00eZ4fcJNK9kCaSMmEugoZWcI7NGbZXzqFWqbpRI7NcDP9+WIQ+i9U5vqWsqd/zng4kbuAJ6UuKqIzB0upYrLShfQE3SAck8oaLhJqqq56VfDuASNpJKidV+zq27HfSBmbXnkR/5AK337dc3MXKJypoK/QPMLKUAP5XLPbs+NddJQV7EZXd29DLgp+fRIg3edpKdO7ZErWhv7d+3Kws+e1Y+ypmR2WIVSwVyBEUfgv2C8Ts9gnTF4pNcEY/S2aBicz5Ew2+jdyGNQQ== test@example.com\n"  # pylint: disable=line-too-long